requirements-api.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlmodel
pytest
httpx
passlib[bcrypt]
python-jose[cryptography]
bcrypt<5
email-validator
psycopg2-binary
redis
cloudinary
google-auth


